+++ crates/core/tedge/src/cli/connect/command.rs
@@ -1,4 +1,6 @@
-use crate::{cli::connect::*, command::Command, system_services::*, ConfigError};
+use crate::{
+    cli::connect::jwt_token::*, cli::connect::*, command::Command, system_services::*, ConfigError,
+};
 use rumqttc::QoS::AtLeastOnce;
 use rumqttc::{Event, Incoming, MqttOptions, Outgoing, Packet};
 use std::path::{Path, PathBuf};
@@ -8,11 +10,11 @@ use tedge_config::*;
 use tedge_utils::paths::{create_directories, ok_if_not_found, DraftFile};
 use which::which;
 
-const DEFAULT_HOST: &str = "localhost";
+pub(crate) const DEFAULT_HOST: &str = "localhost";
 const WAIT_FOR_CHECK_SECONDS: u64 = 10;
 const C8Y_CONFIG_FILENAME: &str = "c8y-bridge.conf";
 const AZURE_CONFIG_FILENAME: &str = "az-bridge.conf";
-const RESPONSE_TIMEOUT: Duration = Duration::from_secs(10);
+pub(crate) const RESPONSE_TIMEOUT: Duration = Duration::from_secs(10);
 const MOSQUITTO_RESTART_TIMEOUT_SECONDS: u64 = 5;
 const MQTT_TLS_PORT: u16 = 8883;
 const TEDGE_BRIDGE_CONF_DIR_PATH: &str = "mosquitto-conf";
@@ -579,130 +581,10 @@ fn check_connected_c8y_tenant_as_configured(configured_url: &str, port: u16) {
     match get_connected_c8y_url(port) {
         Ok(url) if url == configured_url => {}
         Ok(url) => println!(
-            "Warning: Connecting to {}, but configured URL is {}.\n",
+            "Warning: Connecting to {}, but the configured URL is {}.\n\
+            The device certificate has to be removed from the former tenant.\n",
             url, configured_url
         ),
         Err(_) => println!("Failed to get the connected tenant URL from Cumulocity.\n"),
-    }
-}
-
-fn get_connected_c8y_url(port: u16) -> Result<String, ConnectError> {
-    const C8Y_TOPIC_BUILTIN_JWT_TOKEN_UPSTREAM: &str = "c8y/s/uat";
-    const C8Y_TOPIC_BUILTIN_JWT_TOKEN_DOWNSTREAM: &str = "c8y/s/dat";
-    const CLIENT_ID: &str = "get_jwt_token_c8y";
-
-    let mut options = MqttOptions::new(CLIENT_ID, DEFAULT_HOST, port);
-    options.set_keep_alive(RESPONSE_TIMEOUT);
-
-    let (mut client, mut connection) = rumqttc::Client::new(options, 10);
-    let mut acknowledged = false;
-
-    client.subscribe(C8Y_TOPIC_BUILTIN_JWT_TOKEN_DOWNSTREAM, AtLeastOnce)?;
-
-    for event in connection.iter() {
-        match event {
-            Ok(Event::Incoming(Packet::SubAck(_))) => {
-                // We are ready to get the response, hence send the request
-                client.publish(C8Y_TOPIC_BUILTIN_JWT_TOKEN_UPSTREAM, AtLeastOnce, false, "")?;
-            }
-            Ok(Event::Incoming(Packet::PubAck(_))) => {
-                // The request has been sent
-                acknowledged = true;
-            }
-            Ok(Event::Incoming(Packet::Publish(response))) => {
-                // We got a response
-                let token = String::from_utf8(response.payload.to_vec()).unwrap();
-                let connected_url = decode_jwt_token(token.as_str())?;
-                return Ok(connected_url);
-            }
-            Ok(Event::Outgoing(Outgoing::PingReq)) => {
-                // No messages have been received for a while
-                println!("Local MQTT publish has timed out.");
-                break;
-            }
-            Ok(Event::Incoming(Incoming::Disconnect)) => {
-                eprintln!("ERROR: Disconnected");
-                break;
-            }
-            Err(err) => {
-                eprintln!("ERROR: {:?}", err);
-                break;
-            }
-            _ => {}
-        }
-    }
-
-    if acknowledged {
-        // The request has been sent but without a response
-        println!("\nThe request has been sent, however, no response.");
-        Err(ConnectError::TimeoutElapsedError)
-    } else {
-        // The request has not even been sent
-        println!("\nMake sure mosquitto is running.");
-        Err(ConnectError::TimeoutElapsedError)
-    }
-}
-
-fn decode_jwt_token(token: &str) -> Result<String, ConnectError> {
-    // JWT token format: <header>.<payload>.<signature>. Thus, we want only <paylaod>.
-    let payload = token
-        .split_terminator('.')
-        .nth(1)
-        .ok_or(ConnectError::InvalidJWTToken {
-            token: token.to_string(),
-        })?;
-
-    let decoded = base64::decode(payload)?;
-    let json: serde_json::Value = serde_json::from_slice(decoded.as_slice())?;
-
-    let tenant_url = json["iss"].as_str().ok_or(ConnectError::InvalidJWTToken {
-        token: token.to_string(),
-    })?;
-
-    Ok(tenant_url.to_string())
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-    use test_case::test_case;
-
-    #[test]
-    fn check_decode_valid_jwt_token() {
-        let token = "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJqdGkiOm51bGwsImlzcyI6InRlc3QuY3VtdWxvY2l0eS5jb20iLCJhdWQiOiJ0ZXN0LmN1bXVsb2NpdHkuY29tIiwic3ViIjoiZGV2aWNlX3Rlc3QwMDA1IiwidGNpIjoiZGV2aWNlX3Rva2VuX2NvbmZpZyIsImlhdCI6MTYzODQ0Mjk5NywibmJmIjoxNjM4NDQyOTk3LCJleHAiOjE2Mzg0NDY1OTcsInRmYSI6ZmFsc2UsInRlbiI6InQzMTcwNDgiLCJ4c3JmVG9rZW4iOiJLc2VBVUZBTGF1aUplZFFNR2ZzRiJ9.JUYtU9FVWlOWUPJXawFzKNiHD4HoEEWmvKdU1k9L2UF2ofRA2zAdcLH4mxaaspt4suyyZbPL6cS6c9MROG3YCsnqle2NSoYw8mxqncFECWsDS8lwCRTG4402iPTETfWpo9uXw2pFryBoJMAvNzt1qsXXn8EXSYxjzgj0YyxSANypm7PL1kMaprdLuUML_9Cwxf7Z6CRyWkZWWmnQ3lYgV5KMGW7HznkkqcmUCvuXKrHhVL5RkmzE1WyL4ndpGEPFEv9VYmEvFYA8wVHSuw5iVZIFp5lQldDdy_8U-N80xnf3fqZ6Q_wnVm8cga77vIgcf9zK5rSCdehvolM48uM4_w";
-        let expected_url = "test.cumulocity.com";
-        assert_eq!(decode_jwt_token(token).unwrap(), expected_url.to_string());
-    }
-
-    #[test_case(
-    "dGVzdC5jdW11bG9jaXR5LmNvbQ=="
-    ; "not jwt token"
-    )]
-    #[test_case(
-    "aaa.bbb.ccc"
-    ; "payload is not base64 encoded"
-    )]
-    #[test_case(
-    "aaa.dGVzdC5jdW11bG9jaXR5LmNvbQ==.ccc"
-    ; "payload is not json"
-    )]
-    #[test_case(
-    "aaa.eyJqdGkiOm51bGwsImF1ZCI6InRlc3QuY3VtdWxvY2l0eS5jb20iLCJzdWIiOiJkZXZpY2VfdGVzdDAwMDUiLCJ0Y2kiOiJkZXZpY2VfdG9rZW5fY29uZmlnIiwiaWF0IjoxNjM4NDQyOTk3LCJuYmYiOjE2Mzg0NDI5OTcsImV4cCI6MTYzODQ0NjU5NywidGZhIjpmYWxzZSwidGVuIjoidDMxNzA0OCIsInhzcmZUb2tlbiI6IktzZUFVRkFMYXVpSmVkUU1HZnNGIn0=.ccc"
-    ; "payload is json but not contains iss field"
-    )]
-    fn check_decode_invalid_jwt_token(input: &str) {
-        match decode_jwt_token(input) {
-            Ok(_) => panic!("This test should fail"),
-            Err(ConnectError::InvalidJWTToken { token }) => {
-                assert_eq!(input, token)
-            }
-            Err(ConnectError::FromBase64Decode(_)) => {
-                assert_eq!(input, "aaa.bbb.ccc")
-            }
-            Err(ConnectError::FromSerdeJson(_)) => {
-                assert_eq!(input, "aaa.dGVzdC5jdW11bG9jaXR5LmNvbQ==.ccc");
-            }
-            _ => panic!("Unexpected error"),
-        }
     }
 }
