+++ crates/core/tedge/src/cli/connect/error.rs
@@ -39,12 +39,8 @@ pub enum ConnectError {
     #[error("Device is not connected to {cloud} cloud")]
     DeviceNotConnected { cloud: String },
 
-    #[error("Provided JWT token is invalid: {token}")]
-    InvalidJWTToken { token: String },
-
-    #[error("Provided payload is not Base64 encoded.\n{0}")]
-    FromBase64Decode(#[from] base64::DecodeError),
-
-    #[error("Provided JSON does not contain URL.\n{0}")]
-    FromSerdeJson(#[from] serde_json::Error),
+    #[error(
+        "The JWT token received from Cumulocity is invalid.\nToken: {token}\nReason: {reason}"
+    )]
+    InvalidJWTToken { token: String, reason: String },
 }
