+++ crates/core/tedge/src/cli/connect/mod.rs
@@ -10,3 +10,4 @@ mod cli;
 mod command;
 mod common_mosquitto_config;
 mod error;
+mod jwt_token;
